requirements.txt
Faker==9.3.1
mysql-connector-python==8.0.26
protobuf==3.18.1
psycopg2==2.9.1
python-dateutil==2.8.2
python-dotenv==0.19.1
six==1.16.0
text-unidecode==1.3
passlib==1.7.4; platform_system == "Windows"